requirements.txt
# Django stack
Django==4.2.17
dj-database-url==2.3.0
django-storages==1.14.5
whitenoise==6.6.0
django-allauth==65.3.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.1
django-unfold==0.46.0
django-jazzmin==3.0.1

# Authentication & Security
PyJWT==2.9.0
oauthlib==3.2.2
cryptography==44.0.0
python-decouple==3.8

# PDF & DOCX
Pillow==9.5.0
python-docx==1.1.2
reportlab==4.3.1
pymupdf==1.22.3
docx2pdf==0.1.8
unoserver==1.4.0 

# Google Cloud
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-storage==3.0.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
google-api-python-client==2.163.0

# Networking & Requests
requests==2.32.3
urllib3==2.2.3
charset-normalizer==3.4.1
idna==3.10
certifi==2024.12.14

# DB Adapters
psycopg2-binary==2.9.10
mysqlclient==2.2.4
pymysql==1.1.0

# Compiler / Dep support
cffi==1.17.1
pycparser==2.22
typing_extensions==4.12.2

# Misc
gunicorn==21.2.0
sqlparse==0.5.3
bleach==6.1.0
